requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==4.1
django-cors-headers==3.13.0
django-environ==0.9.0
django-stubs==1.12.0
django-stubs-ext==0.5.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
google-trans-new==1.1.9
gTTS==2.2.4
idna==3.3
kombu==5.2.4
mypy==0.971
mypy-extensions==0.4.3
packaging==21.3
prompt-toolkit==3.0.30
PyJWT==2.4.0
PyMySQL==1.0.2
pyparsing==3.0.9
pytz==2022.1
redis==4.3.4
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
types-pytz==2022.1.2
types-PyYAML==6.0.11
typing_extensions==4.3.0
urllib3==1.26.11
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
